requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
anthropic>=0.25.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
python-telegram-bot>=20.0
